refactor: rename scoreboard.js to star.js
